+++ srcs/output/ft_print_int_bits.c
@@ -6,11 +6,12 @@
 /*   By: ffoissey <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/16 13:15:59 by ffoissey          #+#    #+#             */
-/*   Updated: 2019/01/16 13:16:01 by ffoissey         ###   ########.fr       */
+/*   Updated: 2020/04/22 17:15:26 by ffoissey         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
+#include "ft_output.h"
 
 int		ft_print_int_bits(int m)
 {
@@ -27,5 +28,5 @@ int		ft_print_int_bits(int m)
 			write(1, " ", 1);
 		i--;
 	}
-	return (write(1, "\n", 1));
+	return ((int)write(1, "\n", 1));
 }
