+++ srcs/vector/ft_vct_splitchr.c
@@ -6,7 +6,7 @@
 /*   By: ffoissey <ffoisssey@student.42.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/11 14:23:53 by ffoissey          #+#    #+#             */
-/*   Updated: 2020/04/10 18:08:24 by ffoissey         ###   ########.fr       */
+/*   Updated: 2020/04/22 17:21:36 by ffoissey         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ t_vector	*vct_splitchr(t_vector *vct, char c, int flag)
 	{
 		if (vct_len(before) == vct_len(vct))
 			flag = 0;
-		vct_popfrom(vct, vct_len(before) + flag);
+		vct_popfrom(vct, vct_len(before) + (size_t)flag);
 	}
 	return (before);
 }
